refactor: reorganise structure, remove redundant `app`
- remove the unnecessary `app` directory as per feedback
- rename `admin` folder to `manager` to avoid conflict with Django's default admin namespace
- update all relevant file paths and imports to reflect the new directory name
- follow the sub-app approach by creating a sub-app within `api` using `django-admin startapp`
- integrate the new sub-app in `INSTALLED_APPS` to ensure proper functionality
