+++ server/api/manager/admin.py
