+++ server/api/manager/views.py
