+++ server/api/settings.py
@@ -52,6 +52,7 @@ INSTALLED_APPS = [
     "rest_framework",
     "corsheaders",
     "api.healthcheck",
+    "api.manager",
 ]
 
 MIDDLEWARE = [
